Fixes code quality issues
