+++ client/src/model/backend/DTAssets.ts
@@ -10,6 +10,38 @@ import {
 } from 'model/backend/interfaces/sharedInterfaces';
 import FileHandler from 'model/backend/fileHandler';
 
+type CreateFileInput =
+  | FileState
+  | {
+      name: string;
+      content: string;
+      isNew: boolean;
+      isFromCommonLibrary: boolean;
+    };
+
+function filterNewFiles(
+  files: CreateFileInput[] | FileState[],
+): CreateFileInput[] {
+  return (files as CreateFileInput[]).filter(
+    (file): file is CreateFileInput => file.isNew,
+  );
+}
+
+function resolveFilePath(
+  file: CreateFileInput,
+  mainFolderPath: string,
+  lifecycleFolderPath: string,
+): string {
+  const fileType = (file as FileState).type;
+  const mainPath = file.isFromCommonLibrary
+    ? `${mainFolderPath}/common`
+    : mainFolderPath;
+  const lifecyclePath = file.isFromCommonLibrary
+    ? `${mainPath}/lifecycle`
+    : lifecycleFolderPath;
+  return fileType === FileType.LIFECYCLE ? lifecyclePath : mainPath;
+}
+
 export function getFilePath(
   file: FileState,
   mainFolderPath: string,
@@ -45,76 +77,29 @@ class DTAssets implements DTAssetsInterface {
     mainFolderPath: string,
     lifecycleFolderPath: string,
   ): CommitAction[] {
-    const newFiles = (
-      files as Array<
-        | FileState
-        | {
-            name: string;
-            content: string;
-            isNew: boolean;
-            isFromCommonLibrary: boolean;
-          }
-      >
-    ).filter(
-      (
-        file,
-      ): file is
-        | FileState
-        | {
-            name: string;
-            content: string;
-            isNew: boolean;
-            isFromCommonLibrary: boolean;
-          } => file.isNew,
-    );
+    return filterNewFiles(files).map((file) => ({
+      action: 'create' as const,
+      filePath: `${resolveFilePath(file, mainFolderPath, lifecycleFolderPath)}/${file.name}`,
+      content: file.content,
+    }));
+  }
 
-    return newFiles.map((file) => {
-      const fileType = (file as FileState).type;
-      const mainFolderPathUpdated = file.isFromCommonLibrary
-        ? `${mainFolderPath}/common`
-        : mainFolderPath;
-      const lifecycleFolderPathUpdated = file.isFromCommonLibrary
-        ? `${mainFolderPathUpdated}/lifecycle`
-        : lifecycleFolderPath;
-      const filePath =
-        fileType === FileType.LIFECYCLE
-          ? lifecycleFolderPathUpdated
-          : mainFolderPathUpdated;
-      return {
-        action: 'create' as const,
-        filePath: `${filePath}/${file.name}`,
-        content: file.content,
-      };
-    });
+  private buildTriggerContent(): string {
+    return `\ntrigger_${this.DTName}:\n  stage: triggers\n  trigger:\n    include: digital_twins/${this.DTName}/.gitlab-ci.yml\n  rules:\n    - if: '$DTName == "${this.DTName}"'\n      when: always\n  variables:\n    RunnerTag: $RunnerTag\n`;
   }
 
   async buildTriggerAction(): Promise<CommitAction | null> {
     const filePath = `.gitlab-ci.yml`;
     const fileContent = await this.fileHandler.getFileContent(filePath);
 
-    const triggerKey = `trigger_${this.DTName}`;
-    if (fileContent.includes(triggerKey)) {
+    if (fileContent.includes(`trigger_${this.DTName}`)) {
       return null;
     }
-
-    const triggerContent = `
-${triggerKey}:
-  stage: triggers
-  trigger:
-    include: digital_twins/${this.DTName}/.gitlab-ci.yml
-  rules:
-    - if: '$DTName == "${this.DTName}"'
-      when: always
-  variables:
-    RunnerTag: $RunnerTag
-`;
-
-    const updatedContent = `${fileContent.trimEnd()}\n${triggerContent}`;
 
     return {
       action: 'update' as const,
       filePath,
-      content: updatedContent,
+      content: `${fileContent.trimEnd()}\n${this.buildTriggerContent()}`,
     };
   }
 
@@ -130,42 +115,14 @@ ${triggerKey}:
     mainFolderPath: string,
     lifecycleFolderPath: string,
   ): Promise<void> {
-    const newFiles = (
-      files as Array<
-        | FileState
-        | {
-            name: string;
-            content: string;
-            isNew: boolean;
-            isFromCommonLibrary: boolean;
-          }
-      >
-    ).filter(
-      (
-        file,
-      ): file is
-        | FileState
-        | {
-            name: string;
-            content: string;
-            isNew: boolean;
-            isFromCommonLibrary: boolean;
-          } => file.isNew,
-    );
-
     await Promise.all(
-      newFiles.map(async (file) => {
+      filterNewFiles(files).map(async (file) => {
+        const filePath = resolveFilePath(
+          file,
+          mainFolderPath,
+          lifecycleFolderPath,
+        );
         const fileType = (file as FileState).type;
-        const mainFolderPathUpdated = file.isFromCommonLibrary
-          ? `${mainFolderPath}/common`
-          : mainFolderPath;
-        const lifecycleFolderPathUpdated = file.isFromCommonLibrary
-          ? `${mainFolderPathUpdated}/lifecycle`
-          : lifecycleFolderPath;
-        const filePath =
-          fileType === FileType.LIFECYCLE
-            ? lifecycleFolderPathUpdated
-            : mainFolderPathUpdated;
         const commitMessage = `Add ${file.name} to ${fileType} folder`;
         await this.fileHandler.createFile(file, filePath, commitMessage);
       }),
@@ -234,25 +191,11 @@ ${triggerKey}:
     try {
       const fileContent = await this.fileHandler.getFileContent(filePath);
 
-      const triggerKey = `trigger_${this.DTName}`;
-      if (fileContent.includes(triggerKey)) {
+      if (fileContent.includes(`trigger_${this.DTName}`)) {
         return `Trigger already exists in the pipeline for ${this.DTName}`;
       }
 
-      const triggerContent = `
-${triggerKey}:
-  stage: triggers
-  trigger:
-    include: digital_twins/${this.DTName}/.gitlab-ci.yml
-  rules:
-    - if: '$DTName == "${this.DTName}"'
-      when: always
-  variables:
-    RunnerTag: $RunnerTag
-`;
-
-      const updatedContent = `${fileContent.trimEnd()}\n${triggerContent}`;
-
+      const updatedContent = `${fileContent.trimEnd()}\n${this.buildTriggerContent()}`;
       const commitMessage = `Add trigger for ${this.DTName} to .gitlab-ci.yml`;
       await this.fileHandler.updateFile(
         filePath,
