+++ client/test/unit/route/digitaltwins/editor/sidebarRendering.test.tsx
@@ -197,111 +197,4 @@ describe('SidebarRendering', () => {
       );
     });
   });
-
-  describe('renderFileSection', () => {
-    const testCases = [
-      {
-        description: 'LibraryAsset (public)',
-        asset: mockLibraryAsset,
-        setup: () => {
-          mockLibraryAsset.isPrivate = false;
-        },
-      },
-      { description: 'DigitalTwin', asset: mockDigitalTwin },
-    ];
-
-    testCases.forEach(({ description, asset, setup }) => {
-      it(`should render file section correctly and handle file click - ${description}`, () => {
-        if (setup) setup();
-
-        const handleFileClick = jest
-          .spyOn(SidebarFunctions, 'handleFileClick')
-          .mockImplementation(jest.fn());
-
-        render(
-          <SimpleTreeView>
-            {SidebarRendering.renderFileSection(
-              {
-                label: 'label',
-                filesToRender: ['file'],
-                asset,
-                tab: 'create',
-                files,
-                dispatch,
-              },
-              setters,
-            )}
-          </SimpleTreeView>,
-        );
-
-        assertFileTreeBehavior(handleFileClick);
-      });
-    });
-
-    it('should render with null asset', () => {
-      const handleFileClick = jest
-        .spyOn(SidebarFunctions, 'handleFileClick')
-        .mockImplementation(jest.fn());
-
-      render(
-        <SimpleTreeView>
-          {SidebarRendering.renderFileSection(
-            {
-              label: 'Section',
-              filesToRender: ['item1'],
-              asset: null,
-              tab: 'create',
-              files,
-              dispatch,
-            },
-            setters,
-          )}
-        </SimpleTreeView>,
-      );
-
-      expect(screen.getByText('Section')).toBeInTheDocument();
-      fireEvent.click(screen.getByText('Section'));
-      expect(screen.getByText('item1')).toBeInTheDocument();
-      fireEvent.click(screen.getByText('item1'));
-      expect(handleFileClick).toHaveBeenCalled();
-    });
-
-    it('should pass options with library and libraryFiles to handleFileClick', () => {
-      const handleFileClick = jest
-        .spyOn(SidebarFunctions, 'handleFileClick')
-        .mockImplementation(jest.fn());
-
-      render(
-        <SimpleTreeView>
-          {SidebarRendering.renderFileSection(
-            {
-              label: 'label',
-              filesToRender: ['file'],
-              asset: mockDigitalTwin,
-              tab: 'reconfigure',
-              files,
-              dispatch,
-            },
-            setters,
-            {
-              library: true,
-              libraryFiles: [],
-            },
-          )}
-        </SimpleTreeView>,
-      );
-
-      fireEvent.click(screen.getByText('label'));
-      fireEvent.click(screen.getByText('file'));
-      expect(handleFileClick).toHaveBeenCalledWith(
-        expect.anything(),
-        'reconfigure',
-        setters,
-        expect.objectContaining({
-          library: true,
-          libraryFiles: [],
-        }),
-      );
-    });
-  });
 });
